# openapi.yaml
openapi: 3.0.3
info:
  title: Y TU - AI Translation & Human Tone Authentication API
  version: 1.0.0
  description: |
    # Y TU API Documentation
    
    **Y TU API** is a comprehensive AI-powered solution for businesses using AI agents and automated customer communication systems. It provides seamless translation, human tone analysis, and intelligent response generation to eliminate language barriers and enhance customer engagement.
    
    ## 🌟 Key Features
    
    - **Multi-language Translation**: Support for 100+ languages with automatic language detection
    - **Human Tone Analysis**: AI-powered evaluation of formality, emotion, urgency, intent, and cultural context
    - **Response Generation**: Create natural, human-like responses optimized for customer engagement
    - **All-in-One Processing**: Complete pipeline from translation to response generation in a single API call
    - **Cultural Sensitivity**: Responses adapted to different cultural communication styles
    
    ## 🔐 Authentication
    
    All endpoints (except public endpoints) require API key authentication. Include your API key using one of these methods:
    
    - **Header**: `X-API-Key: your-api-key-here`
    - **Bearer Token**: `Authorization: Bearer your-api-key-here`
    
    Get your API key from the [API Keys Dashboard](https://dashboard.ytu-api.com/keys).
    
    ## 📊 Rate Limiting
    
    Rate limits vary by pricing tier. All authenticated responses include rate limit headers:
    
...
servers:
...
paths:
...
  /analyze-tone:
    post:
      summary: Analyze Human Tone
      description: |
        Perform AI-powered analysis of text for human-like tone, formality, emotion, urgency, intent, and cultural context.
        
        **Analysis Dimensions:**
        - **Tone Score**: Overall human-likeness (0-100)
        - **Formality**: Scale from 1 (casual) to 10 (formal)
        - **Emotion**: positive, neutral, negative, or mixed
        - **Urgency**: low, medium, or high
        - **Intent**: question, complaint, request, feedback, greeting, farewell, or other
        - **Cultural Context**: Detected communication style (e.g., western_direct, asian_indirect)
        
        **Use Cases:**
        - Analyze customer sentiment in support tickets
        - Detect urgency for prioritization
        - Understand customer intent for routing
        - Evaluate cultural communication patterns
      tags:
        - Tone Analysis
      operationId: analyzeTone
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToneAnalysisRequest'
            examples:
              frustrated-customer:
                summary: Frustrated Customer
                value:
                  text: "I'm extremely frustrated with this service! This is the third time I've had to contact support."
                  language: "en"
                  context: "Customer service escalation"
              polite-inquiry:
                summary: Polite Inquiry
                value:
                  text: "Good morning, I was wondering if you could provide information about your return policy."
                  language: "en"
                  context: "E-commerce inquiry"
              urgent-request:
                summary: Urgent Request
                value:
                  text: "My account was charged twice! Please fix this immediately!"
                  language: "en"
                  context: "Billing issue"
      responses:
        '200':
          description: Tone analysis successful
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-RateLimit-Tier:
              $ref: '#/components/headers/X-RateLimit-Tier'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ToneAnalysisResponse'
              examples:
                frustrated-tone:
                  summary: Frustrated Customer Analysis
                  value:
                    success: true
                    data:
                      toneScore: 72
                      formality: 4
                      emotion: "negative"
                      emotionIntensity: "high"
                      urgency: "high"
                      intent: "complaint"
                      culturalContext: "western_direct"
                      detailedAnalysis:
                        naturalness: 85
                        contextRelevance: 90
                        emotionalAppropriateness: 95
                        culturalAppropriateness: 88
                      suggestions:
                        - "Respond with empathy and acknowledgment of frustration"
                        - "Offer immediate solution or escalation"
                        - "Use apologetic and reassuring language"
                        - "Provide specific timeline for resolution"
                      timestamp: '2026-01-31T10:31:00.123Z'
                    metadata:
                      requestId: req_tone_def456
                      timestamp: '2026-01-31T10:31:00.123Z'
                      processingTime: 420
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /generate-response:
    post:
      summary: Generate AI Response
      description: |
        Generate contextually appropriate, human-like responses using advanced AI.
        
        **Generation Process:**
        1. Analyze original message context and intent
        2. Generate multiple response candidates (3-5)
        3. Score each candidate on human-likeness, appropriateness, clarity, and cultural sensitivity
        4. Return the best response with alternatives
        
        **Response Styles:**
        - **professional**: Business-appropriate, polished language
        - **casual**: Relaxed, conversational tone
        - **friendly**: Warm, approachable, and personable
        - **formal**: Very formal and respectful
        - **empathetic**: Understanding, compassionate, and supportive
        - **concise**: Brief and to the point
        
        **Quality Scoring:**
        - Human-likeness: Natural language patterns
        - Appropriateness: Context and intent matching
        - Clarity: Readability and comprehension
        - Cultural sensitivity: Cultural appropriateness
      tags:
        - Response Generation
      operationId: generateResponse
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateResponseRequest'
            examples:
              empathetic-response:
                summary: Empathetic Customer Service
                value:
                  originalText: "I'm extremely frustrated with this service! This is the third time I've had to contact support."
                  context: "Customer service escalation"
                  language: "en"
                  tone: "empathetic"
                  numAlternatives: 3
              damaged-product:
                summary: Product Return Response
                value:
                  originalText: "I need to return this product, it arrived damaged."
                  context: "E-commerce customer service"
                  language: "en"
                  tone: "professional"
                  numAlternatives: 2
      responses:
        '200':
          description: Response generated successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-RateLimit-Tier:
              $ref: '#/components/headers/X-RateLimit-Tier'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GenerateResponseResponse'
              examples:
                generated-response:
                  summary: Empathetic Response Generated
                  value:
                    success: true
                    data:
                      response: "I completely understand your frustration, and I sincerely apologize for the inconvenience you've experienced. Having to contact support multiple times is unacceptable, and I want to make this right for you immediately. Let me personally ensure this issue is resolved today. Could you please provide me with your account details so I can prioritize this?"
                      score: 94
                      alternatives:
                        - text: "I'm truly sorry to hear about your experience. This shouldn't have happened, and I take full responsibility for getting this resolved for you right away. What can I help you with specifically?"
                          score: 89
                        - text: "Thank you for your patience. I can see you've contacted us before about this issue. Let me escalate this to our senior team immediately to ensure we resolve this for you today."
                          score: 85
                      language: "en"
                      responseStyle: "empathetic"
                      metadata:
                        totalCandidates: 5
                        generationTime: 1200
                        model: "gpt-4"
                      timestamp: '2026-01-31T10:32:00.123Z'
                    metadata:
                      requestId: req_gen_ghi789
                      timestamp: '2026-01-31T10:32:00.123Z'
                      processingTime: 1250
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /process:
    post:
      summary: Complete Processing Pipeline (Recommended)
      description: |
        Execute the complete Y TU AI pipeline in a single API call. This is the **recommended endpoint** for most use cases.
        
        **Processing Steps:**
        1. **Language Detection**: Automatically detect source language
        2. **Translation**: Translate to target language (if different from source)
        3. **Tone Analysis**: Analyze human tone, emotion, urgency, and intent
        4. **Response Generation**: Generate optimal human-like response
        
        **Processing Options:**
        - `translate`: Enable translation (default: true if targetLanguage provided)
        - `analyzeTone`: Enable tone analysis (default: true)
        - `generateResponse`: Enable response generation (default: false, set to true if you want a response)
        
        **Benefits:**
        - Single API call for complete workflow
        - Reduced latency compared to multiple calls
        - Automatic context passing between steps
        - Optimized pricing for bundled operations
      tags:
        - Processing
      operationId: processMessage
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessRequest'
            examples:
              full-pipeline:
                summary: Complete Pipeline (Spanish → English)
                value:
                  text: "Estoy muy frustrado con este servicio! Esta es la tercera vez que tengo que contactar soporte."
                  sourceLanguage: "es"
                  targetLanguage: "en"
                  context: "SaaS customer service"
                  options:
                    translate: true
                    analyzeTone: true
                    generateResponse: true
              auto-detect:
                summary: Auto-detect with Response
                value:
                  text: "Je suis très satisfait de votre produit!"
                  targetLanguage: "en"
                  context: "E-commerce feedback"
                  options:
                    translate: true
                    analyzeTone: true
                    generateResponse: true
              tone-only:
                summary: Tone Analysis Only (No Translation)
                value:
                  text: "I love this product! It's amazing!"
                  targetLanguage: "en"
                  context: "Product review"
                  options:
                    translate: false
                    analyzeTone: true
                    generateResponse: false
      responses:
        '200':
          description: Processing completed successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-RateLimit-Tier:
              $ref: '#/components/headers/X-RateLimit-Tier'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProcessResponse'
              examples:
                complete-processing:
                  summary: Complete Pipeline Result
                  value:
                    success: true
                    data:
                      originalText: "Estoy muy frustrado con este servicio!"
                      translation:
                        originalText: "Estoy muy frustrado con este servicio!"
                        translatedText: "I'm very frustrated with this service!"
                        sourceLanguage: "es"
                        targetLanguage: "en"
                        confidence: 0.97
                      toneAnalysis:
                        toneScore: 68
                        formality: 3
                        emotion: "negative"
                        emotionIntensity: "high"
                        urgency: "high"
                        intent: "complaint"
                        culturalContext: "western_direct"
                        detailedAnalysis:
                          naturalness: 82
                          contextRelevance: 88
                          emotionalAppropriateness: 92
                          culturalAppropriateness: 85
                        suggestions:
                          - "Acknowledge frustration immediately"
                          - "Provide concrete solution"
                          - "Use empathetic language"
                        timestamp: '2026-01-31T10:33:00.123Z'
                      generatedResponse:
                        response: "I completely understand your frustration, and I sincerely apologize for this experience. Let me personally ensure we resolve this for you right away. What specific issue are you experiencing?"
                        score: 92
                        alternatives:
                          - text: "I'm truly sorry to hear about your frustration. This isn't the experience we want for you. How can I help make this right immediately?"
                            score: 88
                        language: "en"
                        responseStyle: "empathetic"
                        metadata:
                          totalCandidates: 5
                          generationTime: 980
                          model: "gpt-4"
                        timestamp: '2026-01-31T10:33:00.123Z'
                      timestamp: '2026-01-31T10:33:00.123Z'
                    metadata:
                      requestId: req_proc_jkl012
                      timestamp: '2026-01-31T10:33:00.123Z'
                      processingTime: 1850
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key authentication. Include your API key in the `X-API-Key` header.
        
        Example: `X-API-Key: ytu_prod_1234567890abcdef`
        
        Get your API key from the [API Keys Dashboard](https://dashboard.ytu-api.com/keys).

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer token authentication using your API key.
        
        Example: `Authorization: Bearer ytu_prod_1234567890abcdef`
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current time window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current time window
      schema:
        type: integer
        example: 847
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
        example: 1738329600
    X-RateLimit-Tier:
      description: Current pricing tier
      schema:
        type: string
        enum: [FREE, STARTER, PRO, ENTERPRISE]
        example: PRO
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - metadata
      properties:
        success:
          type: boolean
          example: true
          description: Indicates successful request
        data:
          type: object
          description: Response data (varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    ResponseMetadata:
      type: object
      required:
        - requestId
        - timestamp
        - processingTime
      properties:
        requestId:
          type: string
          example: req_1738329600_a1b2c3d4
          description: Unique request identifier for tracking and support
        timestamp:
          type: string
          format: date-time
          example: '2026-01-31T10:30:45.123Z'
          description: ISO 8601 timestamp of the response
        processingTime:
          type: integer
          example: 245
          description: Request processing time in milliseconds
...
    ToneAnalysisRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
          example: "I am extremely disappointed with the service quality."
          description: Text to analyze (1-5000 characters)
        language:
          type: string
          pattern: '^[a-z]{2}$'
          example: "en"
          description: Language code (ISO 639-1). Auto-detected if not provided
        context:
          type: string
          maxLength: 1000
          example: "Customer service response"
          description: Business context for improved analysis accuracy
    ToneAnalysisResponse:
      type: object
      required:
        - toneScore
        - formality
        - emotion
        - emotionIntensity
        - urgency
        - intent
        - culturalContext
        - detailedAnalysis
        - suggestions
        - timestamp
      properties:
        toneScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 85
          description: Overall human-likeness score (0-100, higher is more human-like)
        formality:
          type: integer
          minimum: 1
          maximum: 10
          example: 7
          description: Formality level (1=very casual, 10=very formal)
        emotion:
          type: string
          enum: [positive, neutral, negative, mixed]
          example: "negative"
          description: Primary emotion detected
        emotionIntensity:
          type: string
          enum: [low, medium, high]
          example: "high"
          description: Intensity of the detected emotion
        urgency:
          type: string
          enum: [low, medium, high]
          example: "high"
          description: Urgency level of the message
        intent:
          type: string
          enum: [question, complaint, request, feedback, greeting, farewell, other]
          example: "complaint"
          description: Primary intent of the message
        culturalContext:
          type: string
          example: "western_direct"
          description: Detected cultural communication style
        detailedAnalysis:
          type: object
          required:
            - naturalness
            - contextRelevance
            - emotionalAppropriateness
            - culturalAppropriateness
          properties:
            naturalness:
              type: integer
              minimum: 0
              maximum: 100
              example: 90
              description: Natural language pattern score
            contextRelevance:
              type: integer
              minimum: 0
              maximum: 100
              example: 85
              description: Context relevance score
            emotionalAppropriateness:
              type: integer
              minimum: 0
              maximum: 100
              example: 80
              description: Emotional appropriateness score
            culturalAppropriateness:
              type: integer
              minimum: 0
              maximum: 100
              example: 85
              description: Cultural appropriateness score
        suggestions:
          type: array
          items:
            type: string
          example:
            - "Respond with empathy and acknowledgment"
            - "Address the concern directly"
            - "Use apologetic language"
          description: AI-generated suggestions for responding
        timestamp:
          type: string
          format: date-time
          example: '2026-01-31T10:31:00.123Z'
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing-api-key:
              summary: Missing API Key
              value:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: "Authentication required: Missing or invalid API Key"
                  details:
                    hint: "Include your API key in the X-API-Key header or as a Bearer token"
                metadata:
                  requestId: req_unauth_123
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 3
            invalid-api-key:
              summary: Invalid API Key
              value:
                success: false
                error:
                  code: INVALID_API_KEY
                  message: "The provided API key is invalid or has been revoked"
                  details:
                    hint: "Check your API key in the dashboard at https://dashboard.ytu-api.com/keys"
                metadata:
                  requestId: req_unauth_456
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 5
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing-required-field:
              summary: Missing Required Field
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: "Validation failed: Missing required fields"
                  details:
                    errors:
                      - field: "text"
                        message: "Field is required"
                        type: "required"
                metadata:
                  requestId: req_val_112
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 5
            invalid-language-code:
              summary: Invalid Language Code
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: "Validation failed: Invalid language code"
                  details:
                    errors:
                      - field: "targetLanguage"
                        message: "Must be a valid ISO 639-1 language code (2 lowercase letters)"
                        value: "ENG"
                        type: "pattern"
                metadata:
                  requestId: req_val_113
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 4
            text-too-long:
              summary: Text Exceeds Maximum Length
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: "Validation failed: Text exceeds maximum length"
                  details:
                    errors:
                      - field: "text"
                        message: "Text must not exceed 5000 characters"
                        value: 6543
                        type: "maxLength"
                metadata:
                  requestId: req_val_114
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 3
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal-error:
              summary: Internal Server Error
              value:
                success: false
                error:
                  code: INTERNAL_ERROR
                  message: "An unexpected error occurred. Our team has been notified."
                  details:
                    supportContact: "support@ytu-api.com"
                    statusPage: "https://status.ytu-api.com"
                metadata:
                  requestId: req_500_117
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 150
            ai-service-error:
              summary: AI Service Unavailable
              value:
                success: false
                error:
                  code: SERVICE_UNAVAILABLE
                  message: "AI service temporarily unavailable. Please try again."
                  details:
                    service: "tone-analysis"
                    retryAfter: 30
                metadata:
                  requestId: req_503_118
                  timestamp: '2026-01-31T10:30:45.123Z'
                  processingTime: 5000
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...